+++ src/containerapp/azext_containerapp/_constants.py
@@ -6,6 +6,7 @@ MANAGED_ENVIRONMENT_TYPE = "managed"
 CONNECTED_ENVIRONMENT_TYPE = "connected"
 MANAGED_ENVIRONMENT_RESOURCE_TYPE = "managedEnvironments"
 CONNECTED_ENVIRONMENT_RESOURCE_TYPE = "connectedEnvironments"
+CUSTOM_LOCATION_RESOURCE_TYPE = "customLocations"
 
 MAXIMUM_SECRET_LENGTH = 20
 MAXIMUM_CONTAINER_APP_NAME_LENGTH = 32
@@ -20,11 +21,13 @@ CONTAINER_APPS_SDK_MODELS = "azext_containerapp._sdk_models"
 LOG_ANALYTICS_RP = "Microsoft.OperationalInsights"
 CONTAINER_APPS_RP = "Microsoft.App"
 SERVICE_LINKER_RP = "Microsoft.ServiceLinker"
+EXTENDED_LOCATION_RP = "Microsoft.ExtendedLocation"
+CONTAINER_APP_EXTENSION_TYPE = "microsoft.app.environment"
 
 MANAGED_CERTIFICATE_RT = "managedCertificates"
 PRIVATE_CERTIFICATE_RT = "certificates"
 
-DEV_SERVICE_LIST = ["kafka", "postgres", "redis", "mariadb"]
+DEV_SERVICE_LIST = ["kafka", "postgres", "redis", "mariadb", "qdrant"]
 
 DEV_KAFKA_IMAGE = 'kafka'
 DEV_KAFKA_SERVICE_TYPE = 'kafka'
@@ -42,6 +45,10 @@ DEV_MARIADB_IMAGE = 'mariadb'
 DEV_MARIADB_SERVICE_TYPE = 'mariadb'
 DEV_MARIADB_CONTAINER_NAME = 'mariadb'
 
+DEV_QDRANT_IMAGE = 'qdrant'
+DEV_QDRANT_SERVICE_TYPE = 'qdrant'
+DEV_QDRANT_CONTAINER_NAME = 'qdrant'
+
 PENDING_STATUS = "Pending"
 SUCCEEDED_STATUS = "Succeeded"
 UPDATING_STATUS = "Updating"
@@ -58,6 +65,7 @@ APPLE_SECRET_SETTING_NAME = "apple-provider-authentication-secret"
 UNAUTHENTICATED_CLIENT_ACTION = ['RedirectToLoginPage', 'AllowAnonymous', 'Return401', 'Return403']
 FORWARD_PROXY_CONVENTION = ['NoProxy', 'Standard', 'Custom']
 CHECK_CERTIFICATE_NAME_AVAILABILITY_TYPE = "Microsoft.App/managedEnvironments/certificates"
+CONNECTED_ENV_CHECK_CERTIFICATE_NAME_AVAILABILITY_TYPE = "Microsoft.App/connectedEnvironments/certificates"
 
 NAME_INVALID = "Invalid"
 NAME_ALREADY_EXISTS = "AlreadyExists"